key-logger/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio[client]==5.9.0
pynput==1.7.7
python-engineio==4.7.1
eventlet==0.33.3 